Further splitting the _user_mgt module into sub modules
